base for refactor expts: remove ArchivalGroup-to-METS methods
