+++ src/DigitalPreservation/Preservation.API/Program.cs
@@ -77,6 +77,7 @@ try
         .AddSingleton<IMetsLoader, S3MetsLoader>()
         .AddSingleton<IMetsParser, MetsParser>()
         .AddSingleton<IMetsManager, MetsManager>()
+        .AddSingleton<IMetsFromArchivalGroup, MetsFromArchivalGroup>()
         .AddSingleton<IMetadataManager, MetadataManager>()
         .AddSingleton<IPremisManager<FileFormatMetadata>, PremisManager>()
         .AddSingleton<IPremisManager<ExifMetadata>, PremisManagerExif>()
