+++ src/DigitalPreservation/Storage.Repository.Common/Mets/MetsManager.cs
@@ -26,115 +26,7 @@ public class MetsManager(
         return Result.FailNotNull<MetsFileWrapper>(writeResult.ErrorCode!, writeResult.ErrorMessage);
     }
 
-    public async Task<Result<MetsFileWrapper>> CreateStandardMets(Uri metsLocation, ArchivalGroup archivalGroup, string? agNameFromDeposit)
-    {
-        var (file, mets) = await GetStandardMets(metsLocation, agNameFromDeposit);
-        
-        AddResourceToMets(mets, archivalGroup.Id!, mets.StructMap[0].Div, archivalGroup);
-        
-        var writeResult = await WriteMets(new FullMets{ Mets = mets, Uri = file });
-        if (writeResult.Success)
-        {
-            return await metsParser.GetMetsFileWrapper(file);
-        }
-        return Result.FailNotNull<MetsFileWrapper>(writeResult.ErrorCode!, writeResult.ErrorMessage);
-    }
-
-
-    /// <summary>
-    /// This builds up the METS file from repository resources, not working files
-    /// This will likely never be used in production
-    /// </summary>
-    /// <param name="mets"></param>
-    /// <param name="archivalGroupUri"></param>
-    /// <param name="div"></param>
-    /// <param name="container"></param>
-    private void AddResourceToMets(DigitalPreservation.XmlGen.Mets.Mets mets, Uri archivalGroupUri, DivType div, Container container)
-    {
-        var agLocalPath = archivalGroupUri.LocalPath;
-        foreach (var childContainer in container.Containers)
-        {
-            DivType? childDirectoryDiv = null;
-            if (container is ArchivalGroup && childContainer.GetSlug() == FolderNames.Objects)
-            {
-                // The objects div should already exist from our template
-                childDirectoryDiv = mets.StructMap[0].Div.Div.Single(d => d.Id == Constants.ObjectsDivId);
-            }
-
-            if (childDirectoryDiv == null)
-            {
-                var localPath = childContainer.Id!.LocalPath.RemoveStart(agLocalPath).RemoveStart("/");
-                var admId = Constants.AdmIdPrefix + localPath;
-                var techId = Constants.TechIdPrefix + localPath;
-                childDirectoryDiv = new DivType
-                {
-                    Type = Constants.DirectoryType,
-                    Label = childContainer.Name,
-                    Id = $"{Constants.PhysIdPrefix}{localPath}",
-                    Admid = { admId }
-                };
-                div.Div.Add(childDirectoryDiv);
-                var reducedPremisForObjectDir = new FileFormatMetadata
-                {
-                    Source = Constants.Mets,
-                    OriginalName = localPath,
-                    StorageLocation = childContainer.Id
-                };
-                mets.AmdSec.Add(GetAmdSecType(reducedPremisForObjectDir, admId, techId));
-            }
-
-            AddResourceToMets(mets, archivalGroupUri, childDirectoryDiv, childContainer);
-        }
-
-        AddBinariesToMets(container.Binaries, agLocalPath, div, mets);
-    }
-
-    private void AddBinariesToMets(List<Binary> binaries, string agLocalPath, DivType div, DigitalPreservation.XmlGen.Mets.Mets mets)
-    {
-        foreach (var binary in binaries)
-        {
-            var localPath = binary.Id!.LocalPath.RemoveStart(agLocalPath).RemoveStart("/");
-            if (MetsUtils.IsMetsFile(localPath!, true))
-            {
-                continue;
-            }
-            var fileId = Constants.FileIdPrefix + localPath;
-            var admId = Constants.AdmIdPrefix + localPath;
-            var techId = Constants.TechIdPrefix + localPath;
-            var childItemDiv = new DivType
-            {
-                Type = Constants.ItemType,
-                Label = binary.Name,
-                Id = $"{Constants.PhysIdPrefix}{localPath}",
-                Fptr = { new DivTypeFptr { Fileid = fileId } }
-            };
-            div.Div.Add(childItemDiv);
-            mets.FileSec.FileGrp[0].File.Add(
-                new FileType
-                {
-                    Id = fileId,
-                    Admid = { admId },
-                    Mimetype = binary.ContentType,
-                    FLocat = {
-                        new FileTypeFLocat
-                        {
-                            Href = localPath, Loctype = FileTypeFLocatLoctype.Url
-                        }
-                    }
-                });
-            var premisFile = new FileFormatMetadata
-            {
-                Source = Constants.Mets,
-                Digest = binary.Digest,
-                Size = binary.Size,
-                OriginalName = localPath,
-                StorageLocation = binary.Id
-            };
-            mets.AmdSec.Add(GetAmdSecType(premisFile, admId, techId));
-        }
-    }
-
-    private async Task<(Uri file, DigitalPreservation.XmlGen.Mets.Mets mets)> GetStandardMets(Uri metsLocation, string? agNameFromDeposit)
+    internal async Task<(Uri file, DigitalPreservation.XmlGen.Mets.Mets mets)> GetStandardMets(Uri metsLocation, string? agNameFromDeposit)
     {
         // might be a file path or an S3 URI
         var fileLocResult = await metsParser.GetRootAndFile(metsLocation);
@@ -432,7 +324,7 @@ public class MetsManager(
         return mets;
     }
     
-    private AmdSecType GetAmdSecType(FileFormatMetadata premisFile, string admId, string techId, string? digiprovId = null, VirusScanMetadata? virusScanMetadata = null)
+    internal AmdSecType GetAmdSecType(FileFormatMetadata premisFile, string admId, string techId, string? digiprovId = null, VirusScanMetadata? virusScanMetadata = null)
     {
         var premis = premisManager.Create(premisFile);
         var xElement = premisManager.GetXmlElement(premis, true);
